добавил много чего,но в основном новое это бустеры

+++ src/main/java/svend/taikon/DataBase/ModelDAO/UserDB.java
@@ -71,7 +71,7 @@ public class UserDB implements DAO<User, UUID> {
                 doc.getString("name"),
                 new LargeNumber(doc.getString("income")),
                 new LargeNumber(doc.getString("balance")),
-                doc.getDouble("incomeMultiplier")
+                doc.getInteger("incomeMultiplier")
         );
     }
